Move command logic to pkg
