+++ cmd/kev/main.go
@@ -16,9 +16,7 @@
 
 package main
 
-import (
-	"github.com/appvia/kube-devx/cmd/kev/cmd"
-)
+import "github.com/appvia/kube-devx/cmd/kev/cmd"
 
 func main() {
 	cmd.Execute()
